# infrastructure/apim/specs/api-marketplace.openapi.yml
openapi: 3.0.3
info:
  title: API Marketplace
  description: >-
    Public API for the HMCTS API Marketplace. Covers account registration and
    sign-in for marketplace users, and the inbound request forms used to ask for
    access to an existing API.
  version: 0.0.0
  contact:
    email: no-reply@hmcts.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
servers:
  - description: Service root
    url: /
...
paths:
...
  /api/requests/access:
    post:
      summary: Request access to an existing API
      description: >-
        Submits a request for access to an API already listed in the marketplace.
        The request is stored and a reference is returned for the requester to
        quote. Notification of the API owner is handled separately and its failure
        does not fail this call.
      operationId: createAccessRequest
      tags:
        - requests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRequest'
      responses:
        '201':
          description: Request stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessRequestResponse'
        '400':
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Enter a valid email address.
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/requests/onboarding:
    post:
      summary: Submit an API onboarding request
      description: >-
        Existing onboarding request endpoint, ported from
        service-marketplace-onboarding. Overlaps with POST /api/requests/access —
        see the repository README for the consolidation decision.
      operationId: createOnboardingRequest
      tags:
        - requests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingRequest'
      responses:
        '201':
          description: Request stored
        '400':
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Signed JWT issued by POST /api/register or POST /api/login, sent as
        `Authorization: Bearer <token>`. Valid for 7 days.
  responses:
    TooManyRequests:
      description: >-
        Rate limit exceeded. Applies to the unauthenticated endpoints, which
        permit 20 requests per 15 minutes per client IP.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Too many attempts. Please try again later.
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Something went wrong. Please try again.
  schemas:
...
    AccessRequest:
      description: Request for access to an API already listed in the marketplace
      type: object
      required:
        - fullName
        - organisation
        - email
        - apiName
        - environment
        - useCase
      properties:
        fullName:
          type: string
          description: Full name of the requester
          example: Colin Greenwood
        organisation:
          type: string
          description: Organisation the requester belongs to
          example: HMCTS
        email:
          type: string
          format: email
          description: Contact email address
          example: colin.greenwood@example.com
        jobTitle:
          type: string
          description: Job title of the requester
          example: Software Engineer
        apiName:
          type: string
          description: Name of the API access is being requested for
          example: RAG Service API
        environment:
          type: string
          description: Target environment for access
          example: sandbox
        callVolume:
          type: string
          description: Expected call volume, free text
          example: 1000/day
        useCase:
          type: string
          description: How the API will be used
          example: Automating hearing result document distribution
    AccessRequestResponse:
      description: Confirmation of a stored access request
      type: object
      required:
        - reference
      properties:
        reference:
          type: string
          description: >-
            Reference for the requester to quote, in the form AR-{year}-{6
            alphanumeric characters}.
          pattern: '^AR-[0-9]{4}-[A-Z0-9]{6}$'
          example: AR-2026-K4M9XZ
    OnboardingRequest:
      description: Request to onboard to an API in the marketplace
      type: object
      required:
        - name
        - organisation
        - email
        - jobTitle
        - apiRequested
        - environment
        - useCase
      properties:
        name:
          type: string
          description: Full name of the requester
          example: Colin Greenwood
        organisation:
          type: string
          description: Organisation the requester belongs to
          example: HMCTS
        email:
          type: string
          format: email
          description: Contact email address
          example: colin.greenwood@example.com
        jobTitle:
          type: string
          description: Job title of the requester
          example: Software Engineer
        phone:
          type: string
          description: Contact phone number
          example: "+447700900000"
        apiRequested:
          type: string
          description: Name of the API being requested
          example: RAG Service API
        environment:
          type: string
          description: Target environment for access
          example: sandbox
        callVolume:
          type: string
          description: Expected call volume, free text
          example: 1000/day
        useCase:
          type: string
          description: How the API will be used
          example: Automating hearing result document distribution
    ErrorResponse:
      description: >-
        Error payload. `error` carries a message suitable for display to the end
        user, because the marketplace front end renders it directly.
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message, safe to display to the user
          example: Something went wrong. Please try again.
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
        traceId:
          type: string
          description: Correlation id for locating this failure in the service logs